+++ main.py
@@ -4,6 +4,7 @@ import sys
 
 'import functions from other files'
 from changeSpeed import increaseSpeed, decreaseSpeed
+from changeVolume import increase_volume, decrease_volume
 '''
 Main file of program. It will take in a wav file as the argument and read the samples
 From there the program will augment those samples in different ways based on the users
@@ -29,25 +30,25 @@ sampleRate = samples[0]
 # sci.write(output, samples[0], returnData)
 
 'we can add more functions here, we can also abstract these functions in other files as well'
-def increase_volume(samples,rate):
-    choice = int(input("Increase volume by a factor of how much?  "))
-    data = []
-    for sample in samples:
-        value = sample * choice
-        data.append(value)
-
-    returnData = numpy.array(data)
-    return returnData
-
-def decrease_volume(samples,rate):
-    choice = int(input("Decrease volume by a factor of how much?  "))
-    data = []
-    for sample in samples:
-        value = sample / choice
-        data.append(value)
-
-    returnData = numpy.array(data)
-    return returnData
+# def increase_volume(samples,rate):
+#     choice = int(input("Increase volume by a factor of how much?  "))
+#     data = []
+#     for sample in samples:
+#         value = sample * choice
+#         data.append(value)
+#
+#     returnData = numpy.array(data)
+#     return returnData
+#
+# def decrease_volume(samples,rate):
+#     choice = int(input("Decrease volume by a factor of how much?  "))
+#     data = []
+#     for sample in samples:
+#         value = sample / choice
+#         data.append(value)
+#
+#     returnData = numpy.array(data)
+#     return returnData
 
 while True:
     print("1. Increase volume by factor")
